#60 문서 총길이 계산에 오차가 있는 문제 수정

+++ app/src/main/java/kr/eg/egiwon/pdfsample/util/DefaultSetting.kt
@@ -6,8 +6,8 @@ import javax.inject.Inject
 
 class DefaultSetting @Inject constructor(context: Context) {
 
-    val defaultDocumentSpacing: Float by lazy {
-        DEFAULT_DOCUMENT_SPACING.pixelToDP(context)  // dp
+    val defaultDocumentSpacing: Int by lazy {
+        DEFAULT_DOCUMENT_SPACING.pixelToDP(context).toInt()  // dp
     }
 
     val defaultOffset: Float by lazy {
